03 Post-Processing/requirements-post-processing.txt
opencv-python~=4.5.2.54
numpy~=1.21.0
Pillow~=8.3.1